creating delete images progress bar, and converting delete camera modal to alert dialog
